281: rename look/locations.js to look/tileLocations.js
